requirements.txt
aiosmtplib==4.0.2
annotated-types==0.7.0
anyio==4.11.0
babel==2.17.0
backrefs==5.9
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fakeredis==2.32.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastapi-mail==1.5.6
flake8==7.3.0
ghp-import==2.1.0
greenlet==3.2.4
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
isort==7.0.0
Jinja2==3.1.6
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.22
mkdocs-material-extensions==1.3.1
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
pyyaml_env_tag==1.1
redis==6.4.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
